+++ src/main/java/less/green/openpudo/rest/dto/user/UserProfile.java
@@ -32,6 +32,6 @@ public class UserProfile {
     private UUID profilePicId;
 
     @Schema(readOnly = true)
-    private Integer packageCount;
+    private Long packageCount;
 
 }
